1.stash2

+++ WorldAR_Java/src/main/java/com/huawei/hiardemo/java/fragment/PrruMapFragment.java
@@ -235,17 +235,18 @@ public class PrruMapFragment extends Fragment {
         });
     }
 
-    public void setNowLocation(int x, int y) {
+    public void setNowLocation(float x, float y) {
         if (mCircleShape == null) {
             mCircleShape = new CircleShape("loc", Color.RED);
             mCircleShape.setRadius(10);
             mCircleShape.setValues(x, y);
             mFloorMap.addShape(mCircleShape, false);
+            
         } else {
             mCircleShape.setValues(x, y);
             if (prruInfoShapes != null && prruInfoShapes.size() > 0) {
                 PrruInfoShape minDistacePrru = DistanceUtil.getMinDistacePrru(Float.valueOf(mScale), 5, prruInfoShapes, x, y);//判断距离Prru位置 如果小于1m进行弹窗
-                Toast.makeText(mContext, "请在附件放置prrux="+x+"   ,   "+y, Toast.LENGTH_SHORT).show();
+                Toast.makeText(mContext, "请在附件放置prrux=" + x + "   ,   " + y, Toast.LENGTH_SHORT).show();
                 if (minDistacePrru != null) {
                     Toast.makeText(mContext, "请在附件放置prru", Toast.LENGTH_SHORT).show();
                 }
